Client/client/requirements.txt
fastapi==0.96.0
matplotlib==3.7.1
pandas==2.0.2
pydantic==1.10.9
pytest==7.1.3
python-dotenv==1.0.0
requests==2.31.0
starlette==0.27.0
python-multipart==0.0.6
uvicorn[standard]==0.22.0
pika==1.3.2
streamlit